feat: Logo on Navbar implemented
Wyatt and Cyrus both imported the icon logo and made it as button. Once the button is clicked it will take the user to the landing page. The navbar will follow to any page it goes to as well as it will still stay at the top while the user scrolls down a page.