fix: Fix formatting for LandingPage and CreatePost
Updated the CreatePost page so that the text fields are centered and moved down from the top. Same was done to the PostPage so that the navbar has space. The PostPage is added to be apart of the LandingPage. Routing is fixed in App.js so all pages can be accessed.

#41