enhancement: Create Creation component created
Ari and I implemented and created CaptionCreation.js which includes a function called CreateCaption. When this is called on the individual post page a text box and submission button will appear. Once clicked, the user will get an alert with a response if the caption got posed successfully or not.