enhancement: Implemented Rating Component
Ari and I have implemented a rating component in IndividualPost.js. Upon loading into the page the user will be able to see the captions below the post and a five star rating component next to it. After hovering over it the user can select a rating for the caption. After the on click it will send a post request which will return an alert stating if the request went through or not.