enhancement: Multiple page layouts adjusted
Ari and I have adjust the layouts for the Post , PostList , CaptionCreation , IndividualPost ,and CreatePost files. They are adjusted so that when the window sizes are different the components will scale to fit the screen without any components meshing in to each other. We also add a hover feature on the landing page where a small text box will appear stating 'Click to view post' for user assistance.