enhancement: Added mock up login functionality
Ari and I made a mock up of the log in features. What will happen is depending if the user is logged in or not they will have some features available. We made some mock up boolean state values if a user is logged in or not and implemented the features. This is a mock up and a sandbox branch, will not be put up for a pull request.