enhancement: Implemented localStorage validation
Using local storage we check to see if a token is not null. If it is not null(user logged in) we have a conditional statement that sets a boolean variable called visibility to either true or false. This now shows the correct components on screen if the user is logged in or not. This is a sandbox branch and will not be put up for a pull request.