enhancement: Added more login features and handling 401 errors.
Ari and I created more boolean variables that are set equal to true or false based on if the user is logged in or not. Also added conditional statements to post requests for Ratings, Caption Creation, and Creating a post. With this statement it will redirect the user if there is a 401 authorization error. It will redirect them to the log in page.