enhancement: removed image url textfield
Ari and I removed the image url text field and replaced it with an upload image button that will open up the users finder / file explorer and give them the option to upload an image. Also removed post.imageURL from postlist.js and post.js with a new api call as that is now outdated.